requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.0.1
pdf2docx==0.5.6
PyMuPDF==1.23.7
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
Pillow==10.1.0
gunicorn==21.2.0 